Refactor project structure: update main entry point, add OCI CLI check, and remove unused test files
